Remove extra themes and consolidate utilities
