+++ packages/components/src/components/Link/Link.tsx
@@ -1,6 +1,6 @@
 import {ElementType, createElement, AllHTMLAttributes} from 'react';
 
-import {classNames} from '../utilities/css';
+import classNames from '../../utilities/classNames';
 
 import {variant, linkStyle} from './Link.css';
