+++ packages/components/src/components/Text/Text.css.ts
@@ -1,6 +1,6 @@
 import {styleVariants} from '@vanilla-extract/css';
 
-import {vars} from '../../themes/vars.css';
-import {mapToProperty} from '../../themes/utils';
+import {vars} from '../../theme/vars.css';
+import mapToProperty from '../../utilities/mapToProperty';
 
 export const font = styleVariants(vars.fonts, mapToProperty('fontFamily'));
