+++ packages/components/src/components/index.ts
@@ -8,5 +8,3 @@ export * from './Inline';
 export * from './Link';
 export * from './Stack';
 export * from './Text';
-
-export {classNames} from './utilities/css';
