+++ packages/components/src/theme/index.ts
@@ -1,0 +1,1 @@
+export {vars, themeClass} from './vars.css';
