+++ packages/components/src/theme/vars.css.ts
@@ -1,0 +1,5 @@
+import {createTheme, createThemeContract} from '@vanilla-extract/css';
+import * as tokens from '@polaris/tokens';
+
+export const vars = createThemeContract(tokens);
+export const themeClass = createTheme(vars, tokens);
