--- packages/components/src/themes/admin/index.ts
@@ -1,2 +1,0 @@
-export * from './admin.css';
-export {default as tokens} from './tokens';
