--- packages/components/src/themes/admin/tokens.ts
@@ -1,7 +1,0 @@
-import baseTokens from '../docs/tokens';
-
-const adminTokens = {
-  ...baseTokens,
-};
-
-export default adminTokens;
