--- packages/components/src/themes/docs/docs.css.ts
@@ -1,7 +1,0 @@
-import {createTheme} from '@vanilla-extract/css';
-
-import {vars} from '../vars.css';
-
-import tokens from './tokens';
-
-export const themeClass = createTheme(vars, tokens);
