--- packages/components/src/themes/docs/index.ts
@@ -1,2 +1,0 @@
-export * from './docs.css';
-export {default as tokens} from './tokens';
