--- packages/components/src/themes/docs/tokens.ts
@@ -1,10 +1,0 @@
-import {fonts, fontSizes, fontWeights, spacing} from '@polaris/tokens';
-
-const docsTokens = {
-  fonts,
-  fontSizes,
-  fontWeights,
-  spacing,
-};
-
-export default docsTokens;
