--- packages/components/src/themes/index.ts
@@ -1,9 +1,0 @@
-// Utilities
-export {mapToProperty} from './utils';
-
-// Themes
-export * as adminTheme from './admin';
-export * as docsTheme from './docs';
-
-// Vars
-export {vars} from './vars.css';
