--- packages/components/src/themes/vars.css.ts
@@ -1,5 +1,0 @@
-import {createThemeContract} from '@vanilla-extract/css';
-
-import tokens from './docs/tokens';
-
-export const vars = createThemeContract(tokens);
