+++ packages/components/src/utilities/classNames.ts
@@ -1,5 +1,7 @@
 type Falsy = boolean | undefined | null | 0;
 
-export function classNames(...classes: (string | Falsy)[]) {
+function classNames(...classes: (string | Falsy)[]) {
   return classes.filter(Boolean).join(' ') || undefined;
 }
+
+export default classNames;
