+++ packages/components/src/utilities/index.ts
@@ -1,0 +1,2 @@
+export {default as classNames} from './classNames';
+export {default as mapToProperty} from './mapToProperty';
