+++ packages/components/src/utilities/mapToProperty.ts
@@ -1,9 +1,11 @@
 import {Properties} from 'csstype';
 
-export const mapToProperty =
+const mapToProperty =
   <T extends keyof Properties<string | number>>(property: T) =>
   (value: string | number) => {
     const styleRule = {[property]: value};
 
     return styleRule;
   };
+
+export default mapToProperty;
