+++ website/components/Layout/Layout.tsx
@@ -1,9 +1,7 @@
 import React, {ComponentProps} from 'react';
-import {docsTheme as theme, Container, classNames} from '@polaris/components';
+import {themeClass, Container, classNames} from '@polaris/components';
 
 import {root} from './Layout.css';
-
-const {themeClass} = theme;
 
 export const Layout = ({
   className,
